pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "openai>=2.9.0",
    "pinecone-client>=6.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.45",
    "stripe>=14.0.1",
    "twilio>=9.8.8",
    "uvicorn>=0.38.0",
]
